raven: widen integer literal carrier to i128 for full 128-bit support
Term::Int/Const::Int/Expr::Int/Tok::Int now carry i128 instead of i64,
letting literals beyond i64::MAX magnitude embed exactly through the whole
pipeline (parse -> lower -> infer/verify -> compile -> run). IntTy's
overflow/assumption bound helpers now return i128 directly (no more
i64-clamping), so i128 and u64 range facts are exact where the old i64
carrier had to drop or clamp them. The VM's native Value::Int is now i128
too (codegen's sub-word narrowing generalizes from a 64-bit to a 128-bit
native word), so i128/u128 wrapping arithmetic executes correctly instead
of only being "not narrowed" at the old 64-bit word.

Representation for unsigned 128-bit literals: parsed as u128, then
bit-reinterpreted into the i128 Term::Int carrier (magnitudes above
i128::MAX land as a negative i128 bit pattern).

Boundary (documented in rv-core's IntTy/Term doc comments and exercised by
rv-driver's new i128_tests): u128 magnitudes above i128::MAX parse but
fail their built-in `>= 0` range obligation (sound rejection, not silent
truncation) since that check uses signed i128 comparison - so u128 is
supported end-to-end for 0..=i128::MAX. Separately, *checked* (non-
wrapping) arithmetic on an i128/u128 destination is sound-but-incomplete
right at the true boundary, because proving the result stays within
[i128::MIN, i128::MAX] requires normalizing a comparison against those
exact values, which can itself overflow the trusted solver's own
i128-exact rational arithmetic; wrapping_* sidesteps this and is now the
recommended way to do 128-bit arithmetic that must verify. To avoid
needlessly hitting that solver limitation, range assumptions/obligations
that are tautological under the i128 carrier (a bound of exactly
i128::MIN/i128::MAX) are now omitted rather than asserted.

cargo test --workspace: 513 passed (508 baseline + 5 new), 0 warnings.
All pre-existing .rv examples verify identically to before.

+++ crates/rv-solve/src/tests.rs
@@ -17,7 +17,7 @@ fn cmp(op: BinOp, a: Term, b: Term) -> Prop {
 fn var(x: Sym) -> Term {
     Term::Var(x)
 }
-fn int(n: i64) -> Term {
+fn int(n: i128) -> Term {
     Term::Int(n)
 }
